SceneRenderInfo replaced with more concise RenderScene

+++ Sandbox/Source/main.cpp
@@ -11,7 +11,7 @@ int32_t main(int32_t argc, char** argv)
     settings.GameVersion = RIGEL_MAKE_VERSION(0, 0, 1);
     settings.TargetFPS = 165;
     settings.WindowTitle = "Sandbox";
-    settings.WindowSize = glm::vec2(1600, 900);
+    settings.WindowSize = glm::uvec2(1600, 900);
     settings.AssetManagerThreadPoolSize = 8;
 
     const auto engine = Rigel::Engine::CreateInstance();
